backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# HTTP Requests
httpx==0.26.0
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.17.2
webdriver-manager==4.0.1

# NLP and ML Libraries
nltk==3.8.1
vaderSentiment==3.3.2
scikit-learn==1.4.0
spacy==3.7.2
numpy==1.26.3
pandas==2.2.0

# Google Places API
googlemaps==4.10.0

# CORS
fastapi-cors==0.0.6

# Background Jobs & Task Queue
celery==5.3.4
redis==5.0.1

# Reddit API
praw==7.7.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

